Refactor query filters and edit history modules
