+++ db/records.py
@@ -2,141 +2,12 @@ import logging
 
 logger = logging.getLogger(__name__)
 import datetime
-from db.database import get_connection, SUPPORTS_REGEX
+from db.database import get_connection
 from db.schema import get_field_schema
-from db.validation import validate_table, validate_fields, validate_field
+from db.validation import validate_table, validate_field
+from db.query_filters import _build_filters
 
 
-def _normalize_filter_values(value):
-    """Return a list of non-empty filter values."""
-
-    values = value if isinstance(value, list) else [value]
-    return [v for v in values if v != ""]
-
-
-def _operator_snippet(field: str, value: str, op: str, params: list[str]) -> str:
-    """Build a SQL snippet for a single value based on the operator."""
-
-    if op == "equals":
-        params.append(value)
-        return f"{field} = ?"
-    if op == "starts_with":
-        params.append(f"{value}%")
-        return f"{field} LIKE ?"
-    if op == "ends_with":
-        params.append(f"%{value}")
-        return f"{field} LIKE ?"
-    if op == "not_contains":
-        params.append(f"%{value}%")
-        return f"{field} NOT LIKE ?"
-    if op == "regex":
-        if SUPPORTS_REGEX:
-            params.append(value)
-            return f"{field} REGEXP ?"
-        params.append(f"%{value}%")
-        return f"{field} LIKE ?"
-    # default is "contains"
-    params.append(f"%{value}%")
-    return f"{field} LIKE ?"
-
-
-def _build_field_clause(
-    field: str, values: list[str], op: str, mode: str, params: list[str]
-) -> str | None:
-    """Return a combined clause for all values of a field."""
-
-    snippets = [_operator_snippet(field, v, op, params) for v in values]
-    if not snippets:
-        return None
-    joiner = " AND " if mode == "all" else " OR "
-    return "(" + joiner.join(snippets) + ")"
-
-
-def _apply_date_ranges(
-    date_starts: dict, date_ends: dict, clauses: list[str], params: list[str]
-) -> None:
-    """Append SQL clauses for collected date range filters."""
-
-    for base in set(date_starts) | set(date_ends):
-        start = date_starts.get(base)
-        end = date_ends.get(base)
-        if start and end:
-            clauses.append(f"{base} BETWEEN ? AND ?")
-            params.extend([start, end])
-        elif start:
-            clauses.append(f"{base} >= ?")
-            params.append(start)
-        elif end:
-            clauses.append(f"{base} <= ?")
-            params.append(end)
-
-
-
-def _build_filters(table, search=None, filters=None, ops=None, modes=None):
-    """Return SQL where clauses and params for the provided filters/search."""
-    clauses = []
-    params = []
-
-    if filters:
-        valid_keys = [
-            f[:-4] if f.endswith('_min') or f.endswith('_max') else
-            f[:-6] if f.endswith('_start') else
-            f[:-4] if f.endswith('_end') else f
-            for f in filters.keys()
-        ]
-        validate_fields(table, valid_keys)
-
-        # Hold start/end pieces of date range filters until all values are parsed
-        date_starts = {}
-        date_ends = {}
-
-        for fld, val in filters.items():
-            clean_values = _normalize_filter_values(val)
-            if not clean_values:
-                continue
-            if fld.endswith('_min'):
-                base = fld[:-4]
-                for v in clean_values:
-                    clauses.append(f"{base} >= ?")
-                    params.append(v)
-            elif fld.endswith('_max'):
-                base = fld[:-4]
-                for v in clean_values:
-                    clauses.append(f"{base} <= ?")
-                    params.append(v)
-            elif fld.endswith('_start'):
-                base = fld[:-6]
-                date_starts[base] = clean_values[0]
-            elif fld.endswith('_end'):
-                base = fld[:-4]
-                date_ends[base] = clean_values[0]
-            else:
-                op = (ops or {}).get(fld, "contains")
-                mode = (modes or {}).get(fld, "any")
-                clause = _build_field_clause(fld, clean_values, op, mode, params)
-                if clause:
-                    clauses.append(clause)
-
-        # After processing all filters, turn date range pieces into clauses
-        _apply_date_ranges(date_starts, date_ends, clauses, params)
-
-    if search:
-        # Build a search clause across all text-like fields
-        search_term = search.strip()
-        all_fields = get_field_schema()[table]
-        search_fields = [
-            field
-            for field, meta in all_fields.items()
-            if meta["type"] in ("text", "textarea", "select", "multi_select", "url")
-        ]
-        if search_fields:
-            validate_fields(table, search_fields)
-            # Create "field LIKE" clauses for every searchable field
-            subconds = [f"{f} LIKE ?" for f in search_fields]
-            clauses.append("(" + " OR ".join(subconds) + ")")
-            params.extend([f"%{search_term}%"] * len(subconds))
-
-    return clauses, params
 
 def get_all_records(
     table,
@@ -277,123 +148,6 @@ def update_field_value(table, record_id, field, new_value):
             touch_last_edited(table, record_id)
         return success
 
-
-def append_edit_log(
-    table: str,
-    record_id: int,
-    field_name: str | None,
-    old_value: str | None,
-    new_value: str | None,
-    actor: str | None = None,
-) -> None:
-    """Insert a row into the edit_history table."""
-    validate_table(table)
-
-    timestamp = datetime.datetime.utcnow().strftime("%Y-%m-%dT%H:%M:%S")
-
-    with get_connection() as conn:
-        cursor = conn.cursor()
-        try:
-            logger.debug(
-                "append_edit_log: %s id=%s field=%s old=%r new=%r actor=%r",
-                table,
-                record_id,
-                field_name,
-                old_value,
-                new_value,
-                actor,
-            )
-            cursor.execute(
-                """
-                INSERT INTO edit_history
-                    (table_name, record_id, timestamp, field_name, old_value, new_value, actor)
-                VALUES (?, ?, ?, ?, ?, ?, ?)
-                """,
-                (table, record_id, timestamp, field_name, old_value, new_value, actor),
-            )
-            conn.commit()
-            logger.info(
-                "Logged edit for %s id=%s field=%s old=%r new=%r",
-                table,
-                record_id,
-                field_name,
-                old_value,
-                new_value,
-            )
-        except Exception as e:
-            logger.warning("[EDIT LOG ERROR] %s", e)
-
-
-def get_edit_history(table_name: str, record_id: int, limit: int | None = None) -> list[dict]:
-    """Return edit history rows ordered by timestamp descending."""
-    validate_table(table_name)
-
-    with get_connection() as conn:
-        cur = conn.cursor()
-        sql = (
-            "SELECT id, table_name, record_id, timestamp, field_name, old_value, new_value, actor "
-            "FROM edit_history WHERE table_name = ? AND record_id = ? ORDER BY timestamp DESC"
-        )
-        params = [table_name, record_id]
-        if limit is not None:
-            sql += " LIMIT ?"
-            params.append(limit)
-        cur.execute(sql, params)
-        rows = cur.fetchall()
-        cols = [d[0] for d in cur.description]
-        return [dict(zip(cols, r)) for r in rows]
-
-
-def get_edit_entry(edit_id: int) -> dict | None:
-    """Return a single edit_history row by id."""
-    with get_connection() as conn:
-        cur = conn.cursor()
-        cur.execute(
-            "SELECT id, table_name, record_id, timestamp, field_name, old_value, new_value, actor "
-            "FROM edit_history WHERE id = ?",
-            (edit_id,),
-        )
-        row = cur.fetchone()
-        if row:
-            cols = [d[0] for d in cur.description]
-            return dict(zip(cols, row))
-        return None
-
-
-def revert_edit(entry: dict) -> bool:
-    """Undo the provided edit_history entry."""
-    table = entry["table_name"]
-    record_id = entry["record_id"]
-    field = entry["field_name"]
-    old_val = entry["old_value"]
-    new_val = entry["new_value"]
-
-    try:
-        if not field:
-            logger.warning("revert_edit: no field_name provided")
-            return False
-        if field.startswith("relation_"):
-            from db.relationships import add_relationship, remove_relationship
-
-            rel_table = field[len("relation_") :]
-            if old_val is None and new_val is not None:
-                add_relationship(
-                    table, record_id, rel_table, int(new_val), actor="undo"
-                )
-            elif new_val is None and old_val is not None:
-                remove_relationship(
-                    table, record_id, rel_table, int(old_val), actor="undo"
-                )
-            else:
-                return False
-        else:
-            update_field_value(table, record_id, field, old_val)
-        if not field.startswith("relation_"):
-            append_edit_log(table, record_id, field, new_val, old_val, actor="undo")
-    except Exception:
-        logger.exception("Failed to revert edit")
-        return False
-    return True
 
 def create_record(table, form_data):
     # 1) Validate the table name
